Revert "Split UsersSection into components for better maintainability (#19848)"
This reverts commit ab8366fb7e5d28ce7b3b4c94b5909c329c03a18e.

Fixes #20498
