--- js/apps/admin-ui/src/components/users/UserDataTableToolbarItems.tsx
@@ -1,97 +1,0 @@
-import type RealmRepresentation from "@keycloak/keycloak-admin-client/lib/defs/realmRepresentation";
-import {
-  Button,
-  ButtonVariant,
-  Dropdown,
-  DropdownItem,
-  KebabToggle,
-  ToolbarItem,
-} from "@patternfly/react-core";
-import { useState } from "react";
-import { useTranslation } from "react-i18next";
-import { useAccess } from "../../context/access/Access";
-
-type UserDataTableToolbarItemsProps = {
-  realm: RealmRepresentation;
-  hasSelectedRows: boolean;
-  toggleDeleteDialog: () => void;
-  toggleUnlockUsersDialog: () => void;
-  goToCreate: () => void;
-};
-
-export function UserDataTableToolbarItems({
-  realm,
-  hasSelectedRows,
-  toggleDeleteDialog,
-  toggleUnlockUsersDialog,
-  goToCreate,
-}: UserDataTableToolbarItemsProps) {
-  const { t } = useTranslation("users");
-  const [kebabOpen, setKebabOpen] = useState(false);
-
-  const { hasAccess } = useAccess();
-
-  // Only needs query-users access to attempt add/delete of users.
-  // This is because the user could have fine-grained access to users
-  // of a group.  There is no way to know this without searching the
-  // permissions of every group.
-  const isManager = hasAccess("query-users");
-
-  const bruteForceProtectionToolbarItem = !realm.bruteForceProtected ? (
-    <ToolbarItem>
-      <Button
-        variant={ButtonVariant.link}
-        onClick={toggleDeleteDialog}
-        data-testid="delete-user-btn"
-        isDisabled={hasSelectedRows}
-      >
-        {t("deleteUser")}
-      </Button>
-    </ToolbarItem>
-  ) : (
-    <ToolbarItem>
-      <Dropdown
-        toggle={<KebabToggle onToggle={(open) => setKebabOpen(open)} />}
-        isOpen={kebabOpen}
-        isPlain
-        dropdownItems={[
-          <DropdownItem
-            key="deleteUser"
-            component="button"
-            isDisabled={hasSelectedRows}
-            onClick={() => {
-              toggleDeleteDialog();
-              setKebabOpen(false);
-            }}
-          >
-            {t("deleteUser")}
-          </DropdownItem>,
-
-          <DropdownItem
-            key="unlock"
-            component="button"
-            onClick={() => {
-              toggleUnlockUsersDialog();
-              setKebabOpen(false);
-            }}
-          >
-            {t("unlockAllUsers")}
-          </DropdownItem>,
-        ]}
-      />
-    </ToolbarItem>
-  );
-
-  const actionItems = (
-    <>
-      <ToolbarItem>
-        <Button data-testid="add-user" onClick={goToCreate}>
-          {t("addUser")}
-        </Button>
-      </ToolbarItem>
-      {bruteForceProtectionToolbarItem}
-    </>
-  );
-
-  return isManager ? actionItems : null;
-}
